requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-telegram-bot>=20.0
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
twilio>=8.0.0
